libsec: comparison of integers of different signs
../../../common/acl/acl_common.c:1757:13: error: comparison of integers of different signs: 'uint64_t' (aka 'unsigned long long') and 'int' [-Werror,-Wsign-compare]
        if (cookie >= aclcnt)
            ~~~~~~ ^  ~~~~~~

../common/acltext.c:1268:18: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        if (strlen(str) == table_length) {
            ~~~~~~~~~~~ ^  ~~~~~~~~~~~~
1 error generated.

+++ usr/src/lib/libsec/common/acltext.c
@@ -1265,7 +1265,7 @@ ace_inherit_helper(char *str, uint32_t *imask, int table_length)
 {
 	int rc = 0;
 
-	if (strlen(str) == table_length) {
+	if (strlen(str) == (size_t)table_length) {
 		/*
 		 * If the string == table_length then first check to see it's
 		 * in positional format.  If that fails then see if it's in
